pyproject.toml
[tool.poetry]
name = "nb-cli-plugin-webui"
version = "0.4.2"
description = "web ui support for nb-cli"
authors = ["Kyomotoi <kyomotoiowo@gmail.com>"]
license = "MIT"
packages = [{ include = 'nb_cli_plugin_webui' }]
include = ["nb_cli_plugin_webui/dist/**/*"]
readme = "README.md"
homepage = "https://github.com/nonebot/cli-plugin-webui"
repository = "https://github.com/nonebot/cli-plugin-webui"
documentation = "https://github.com/nonebot/cli-plugin-webui"
keywords = ["nonebot", "cli", "web"]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Framework :: Robot Framework",
  "Framework :: Robot Framework :: Library",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
]

[tool.poetry.dependencies]
python = "^3.8.1"
psutil = "^5.9.5"
pathlib = "^1.0.1"
WMI = { version = "^1.5.1", markers = "sys_platform == 'win32'" }
nb-cli = "^1.2.3"
fastapi = "^0.95.1"
APScheduler = "^3.10.1"
loguru = "^0.7.0"
PyJWT = "^2.7.0"
bcrypt = "^4.0.1"
passlib = "^1.7.4"
uvicorn = { extras = ["standard"], version = "^0.22.0" }
httpx = { extras = ["socks"], version = "^0.24.1" }
python-dotenv = "^1.0.0"
pyzipper = "^0.3.6"

[tool.poetry.group.dev.dependencies]
isort = "^5.10.1"
black = "^23.1.0"
nonemoji = "^0.1.2"
pre-commit = "^3.1.0"
flake8 = "^6.0.0"
babel = "^2.16.0"

[tool.poetry.plugins.nb]
webui = "nb_cli_plugin_webui.plugin:install"

[tool.poetry.scripts]
init = "nb_cli_plugin_webui.scripts:init"
update = "nb_cli_plugin_webui.scripts:update"
compile = "nb_cli_plugin_webui.scripts:compile"
generate = "nb_cli_plugin_webui.scripts:generate_openapi"

[tool.black]
line-length = 88
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 88
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.pycln]
path = "."

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
